Rename activate -> runEffects
